feat: Change order of test
